Переделал первый экран с MVC на MVVM.
