+++ MovieBase/MovieBase/SceneDelegate.swift
@@ -11,7 +11,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
         // guard let _ = (scene as? UIWindowScene) else { return }
-        let viewController = CategoryViewController()
+        let viewController = MainViewController()
         let navViewController = UINavigationController(rootViewController: viewController)
 
         window?.rootViewController = navViewController
